requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
plotly>=5.0.0
rich>=12.0.0
typer>=0.7.0
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-aws>=0.1.0
langgraph>=0.0.20
langgraph-supervisor>=0.0.1
openai>=1.0.0

# Bioinformatics
scanpy>=1.9.3
scipy>=1.10.0
scikit-learn>=1.3.0
anndata>=0.9.0
biopython>=1.81
leidenalg>=0.9.0
igraph>=0.10.4
scrublet>=0.2.3
h5py>=3.9.0
tables>=3.8.0

# Visualization
seaborn>=0.12.0
matplotlib>=3.7.0
kaleido>=0.2.0

# Data I/O
openpyxl>=3.1.0
pyarrow>=12.0.0
pyreadr>=0.4.0

# HTTP and utilities
requests>=2.31.0
aiofiles>=23.0.0
xmltodict>=0.13.0

# API support
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing (basic)
pytest>=7.0.0
